Convert project to Maven.
This PR transitions this project to Maven, the same build tool used by
Dapr Java SDK. This alignment will reduce the number of tools used
by Java repositories in the project and allow for more code re-use.

We define only the major scaffolding for the project, re-using as
much configuration from Dapr Java SDK as sensible. Future PRs will
target the auto-generated from protocol buffers and the SDK proper.

We also moving the required JDK to JDK 17 - which should reduce the need
to add Lombok to the project.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>