Support for user-provided StateStore bulk operations.
Moves bulk state store operations to StateStore interface and define
default implementations for them. This means that users don't **have**
to define anything beyond `set`, `get` and `delete` to have a fully
working State Store **with** bulk-operation support. The default
implementation of bulk operations rely on single-item operation but users
are allowed to redefine some or all of those bulk operations in their
implementation if they so desire.

Closes #15.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>