Address review comments.
Provide a helper method for consumers to iterate and process the
contents of a `TransactionalStateRequest#operations()`.  Since this is
a request type, and not something that would be user-defined or
constructed, this should provide a simple and elegant way for users to
parse and process those lists of operations. It also indirectly
documents the kind of operations that can be part of a transaction.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>