Address threading concerns for bidirectional streams.
The code handling bidirectional streams was delegating the whole stream
processing by the component to either the body the main handler or the
body of a StreamObserver::onNext. Either way, processing of new events
would be blocked.

We delegate the component's stream processing task to a separate thread,
addressing this concern.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>