Stop trying to map on a Mono<Void>.
Some of our APIs use `Mono<Void>` to signal success or error.  `Ping` is
one such API and its default implementation (and that for other similar
methods) use `Mono.empty()` to signal success. This makes sense: that
Mono will emit an `onComplete` signal and that's it. Besides, there
isn't really a sane way emit an item through an `onNext()` on a
`Mono<Void>`.  The only signals such Mono would ever produce would be an
`onCompleted` or an `onError`. Thus, calling `.map` method on them would
not produce any item (as there is no item to map from) and be
meaningless.

But our code was indeed calling `map()` and expecting an item to be
produced from that. That was resuling in an error from gRPC: a
StremObserver had its `onCompleted` invoked but it was expecting a
`onNext` to be called before. As that expectation failed, the gRPC call
failed.

This PR addresses this by converting those `.map` on `Mono<Void>`s to
`.thenReturn`, enforcing the intended behaviour.

Reverts and fix #35.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>