Refactor common tests on aspects to a single class.
We are using or abusing of JUnit5 @TestFactories here to avoid
duplicating code that tests for common "aspects" among wrappers (ping,
getFeatures and init). We also had to abuse lambdas and passing methods
arround to circumvent some limitations with generics and the fact that
all theser services expose methods like Init with their very own
sub-types.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>